+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## WIP
+### Added
+- Exposed `AlgoBalanced1Settings` to disable event/state ID to string functions.
+  - https://github.com/StateSmith/StateSmith/issues/181
+
+---
+
 ## [0.9.1-alpha]
 ### Added
 - Expansions - provide meta info (current state/vertex, trigger, behavior).